git-svn-id: https://italc.svn.sourceforge.net/svnroot/italc@150 d6290f70-d918-0410-89a7-c2321b9b1c3d

+++ ica/win32/src/vncEncoder.cpp
@@ -229,7 +229,7 @@ vncEncoder::GetRemotePalette(RGBQUAD *quadlist, UINT ncolours)
 		vnclog.Print(LL_INTINFO, VNCLOG("generating 8-bit palette data\n"));
 
 		rfbPixelFormat remote;
-		remote.trueColour = TRUE;
+		remote.trueColour = (uint8_t)TRUE;
 		remote.bitsPerPixel = 32;
 		remote.depth = 24;
 		remote.bigEndian = FALSE;
